test passing for an attempt to add empty comments

+++ tests/updateComments_test.py
@@ -10,10 +10,14 @@ def comment_file():
         os.remove(comment_test_file)
 
     yield comment_test_file
-    os.remove(comment_test_file)
+    if os.path.exists(comment_test_file):
+        os.remove(comment_test_file)
 
 
 def check_file_contents(test_file, expected_text):
+    if expected_text == "":
+        return not os.path.exists(test_file)
+
     with open(test_file, 'r') as file:
         file_contents = file.read()
     return file_contents == expected_text
